ARM: remove NWFPE and VFP emulation code from the kernel
It has been brought to my attention that the softfloat license is
incompatible with GPLv2 as it contains additional restrictions, so
this code must be removed from the kernel.

Unfortunately, this renders Linux fairly useless on older ARM CPUs
and any ARM VFP that traps to support code for unhandled instructions,
but alas the legal situation takes precident.

Those using ARMv4 and ARMv5 CPUs will have to find the FASTFPE
alternative to NWFPE and see whether it can be resurected.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>